+++ src/main/java/twilightforest/entity/ai/EntityAITFHoverBeam.java
@@ -149,7 +149,7 @@ public class EntityAITFHoverBeam extends EntityAIBase {
 
         double distanceDesired = offsetX * offsetX + offsetY * offsetY + offsetZ * offsetZ;
 
-        distanceDesired = (double) MathHelper.sqrt_double(distanceDesired);
+        distanceDesired = MathHelper.sqrt_double(distanceDesired);
 
         if (distanceDesired > 0.5) {
 
@@ -199,7 +199,7 @@ public class EntityAITFHoverBeam extends EntityAIBase {
             if (possibleEntity.canBeCollidedWith() && possibleEntity != this.attacker) {
                 float borderSize = possibleEntity.getCollisionBorderSize();
                 AxisAlignedBB collisionBB = possibleEntity.boundingBox
-                        .expand((double) borderSize, (double) borderSize, (double) borderSize);
+                        .expand(borderSize, borderSize, borderSize);
                 MovingObjectPosition interceptPos = collisionBB.calculateIntercept(srcVec, destVec);
 
                 if (collisionBB.isVecInside(srcVec)) {
