+++ src/main/java/twilightforest/entity/boss/EntityTFKnightPhantom.java
@@ -274,7 +274,7 @@ public class EntityTFKnightPhantom extends EntityFlying implements IMob {
 
         double factor = moveX * moveX + moveY * moveY + moveZ * moveZ;
 
-        factor = (double) MathHelper.sqrt_double(factor);
+        factor = MathHelper.sqrt_double(factor);
 
         double speed = 0.1D;// this.isChargingAtPlayer() ? 0.1D : 0.05D;
 
@@ -337,10 +337,8 @@ public class EntityTFKnightPhantom extends EntityFlying implements IMob {
 
         if (flag) {
             if (i > 0) {
-                entity.addVelocity(
-                        (double) (-MathHelper.sin(this.rotationYaw * (float) Math.PI / 180.0F) * (float) i * 0.5F),
-                        0.1D,
-                        (double) (MathHelper.cos(this.rotationYaw * (float) Math.PI / 180.0F) * (float) i * 0.5F));
+                entity.addVelocity(-MathHelper.sin(this.rotationYaw * (float) Math.PI / 180.0F) * (float) i * 0.5F,
+                        0.1D, MathHelper.cos(this.rotationYaw * (float) Math.PI / 180.0F) * (float) i * 0.5F);
                 this.motionX *= 0.6D;
                 this.motionZ *= 0.6D;
             }
@@ -451,7 +449,7 @@ public class EntityTFKnightPhantom extends EntityFlying implements IMob {
         this.motionY /= 2.0D;
         this.motionZ /= 2.0D;
         this.motionX -= par3 / (double) f * (double) distance;
-        this.motionY += (double) distance;
+        this.motionY += distance;
         this.motionZ -= par5 / (double) f * (double) distance;
 
         if (this.motionY > 0.4000000059604645D) {
