+++ src/main/java/twilightforest/item/ItemTFGiantPick.java
@@ -106,7 +106,7 @@ public class ItemTFGiantPick extends ItemPickaxe {
         // add new one
         multimap.put(
                 SharedMonsterAttributes.attackDamage.getAttributeUnlocalizedName(),
-                new AttributeModifier(field_111210_e, "Tool modifier", (double) this.damageVsEntity, 0));
+                new AttributeModifier(field_111210_e, "Tool modifier", this.damageVsEntity, 0));
         return multimap;
     }
