+++ src/main/java/twilightforest/item/ItemTFMinotaurAxe.java
@@ -116,7 +116,7 @@ public class ItemTFMinotaurAxe extends ItemAxe {
         // add new one
         multimap.put(
                 SharedMonsterAttributes.attackDamage.getAttributeUnlocalizedName(),
-                new AttributeModifier(field_111210_e, "Tool modifier", (double) this.damageVsEntity, 0));
+                new AttributeModifier(field_111210_e, "Tool modifier", this.damageVsEntity, 0));
         return multimap;
     }
 }
