+++ src/main/java/twilightforest/tileentity/TileEntityTFGhastTrapActive.java
@@ -139,13 +139,8 @@ public class TileEntityTFGhastTrapActive extends TileEntity {
 
         if (!worldObj.isRemote) {
             // trap nearby ghasts
-            AxisAlignedBB aabb = AxisAlignedBB.getBoundingBox(
-                    (double) this.xCoord,
-                    (double) this.yCoord + 16,
-                    (double) this.zCoord,
-                    (double) (this.xCoord + 1),
-                    (double) (this.yCoord + 16 + 1),
-                    (double) (this.zCoord + 1)).expand(6D, 16D, 6D);
+            AxisAlignedBB aabb = AxisAlignedBB.getBoundingBox(this.xCoord,
+                    (double) this.yCoord + 16, this.zCoord, this.xCoord + 1, this.yCoord + 16 + 1, this.zCoord + 1).expand(6D, 16D, 6D);
 
             List<EntityGhast> nearbyGhasts = worldObj.getEntitiesWithinAABB(EntityGhast.class, aabb);
